Workaround for removing event categories from the 'New Event' page.
We attempt to deregister event category support in the OL theme.  However,
this creates issues when attempting to fetch events via AJAX, which
usually occurs when viewing the event calendar.

This commit implements a workaround that reinstates event category
support, but hides the Categories field from the 'New Event' page via CSS.

We should revert this commit when Event Organiser pushes out a new release
that implements the proper fix and when we're ready to pin the latest
version.

See https://github.com/cuny-academic-commons/commons-in-a-box/issues/163