Improve behavior of new_group_id cookie during group creation navigation.
When visiting /groups/create/step/group-details (step 1 of group creation),
the new_group_id cookie and global should be reset. This ensures that aborted
group creation processes don't interfere the next time you try to create a group.

The exception is when you navigate to group-details using the 'Previous
Step' button, which we detect through the use of a new URL param `is_previous`.

See cuny-academic-commons/commons-in-a-box#436.