Adjust Manage Members element height using JS.
The proper fix is to use CSS grid or a similar technique, but this
is a large refactor. For the time being, we dynamically set element
height to match the largest element.

This is necessitated by the 'Membership Hidden' indicator introduced
in #469, but it was already a problem in other edge cases, such as
when the member name is really long.

See cuny-academic-commons/commons-in-a-box#469.